+++ Taylan/sys-prog/threading/file1.dat
@@ -1,7 +1,7 @@
 test
 test
 test
-ali
-veli
-kirk
-dokuz
+eer
+reee
+ree
+reee
