+++ Taylan/sys-prog/threading/file2.dat
@@ -1,7 +1,6 @@
-veli
-veli
-kirk
-dokuz
-asd
-fdsa
-zxc
+scent
+sense
+scent
+sense
+sensitivity
+sensitive
