+++ Taylan/sys-prog/threading/src/com/kondanta/FileAccessors.java
@@ -6,37 +6,80 @@ import java.io.IOException;
 import java.util.HashMap;
 
 class FileAccessors{
-    private HashMap<String, Integer> occurrenceTable =
-            new HashMap<>();
 
-    private void initTable(String key, Integer counter) {
-        occurrenceTable.putIfAbsent(key, counter);
+    /**
+     * Hash table that is going to store parsed strings and their occurrence
+     * numbers
+     */
+    private static HashMap<String, Integer> occurrenceTable =
+            new HashMap<>();
+    /**
+     * Mutex for multi threaded approach
+     */
+    private final Object lock = new Object();
+
+
+    /**
+     * @param key Hash table key value
+     *            Populates the hash map with the key if its not already existing
+     *            in the hash table.
+     */
+    private void initTable(String key) {
+        occurrenceTable.putIfAbsent(key, 0);
     }
 
-    void replaceValue(String key, Integer oldVal, Integer newVal){
+    /**
+     * @param key    Hash table key that its value is going to be changed.
+     * @param oldVal Old value of the key
+     * @param newVal New value of the key
+     */
+    private void replaceValue(String key, Integer oldVal, Integer newVal) {
         occurrenceTable.replace(key, oldVal, newVal);
     }
 
-    void printer(String word){
-        System.out.printf("The word %s occured %d", word, occurrenceTable.get(word));
+
+    /**
+     * @param key Hash table key that is going to be searched
+     * @return true if key exists in the table
+     */
+    private boolean isExists(String key) {
+        return occurrenceTable.containsKey(key);
     }
 
-    void read(String fileName) {
-        Integer ctr = 0;
-        try (BufferedReader br = new BufferedReader(new FileReader(fileName))) {
-            StringBuilder sb = new StringBuilder();
-            String line = br.readLine();
+    /**
+     * @param key Hash table key
+     * @return Value of the key
+     */
+    private Integer value(String key) {
+        return occurrenceTable.get(key);
+    }
 
-            while (line != null) {
-                sb.append(line);
-                sb.append(System.lineSeparator());
-                line = br.readLine();
-                this.initTable(line, ctr);
-                if(occurrenceTable.containsKey(line)) {
-                    replaceValue(line, ctr, ctr++);
+    /**
+     * @param word Searched value on the table
+     *             Prints the occurrence number of the word.
+     */
+    void printOccurrenceNumber(String word) {
+        System.out.printf("The word %s occurred %d", word, this.value(word));
+    }
+
+
+    /**
+     * @param fileName File that is going to be parsed
+     * Takes fileName and populates the hash table with the words that read from the file.
+     * If the word exists in table, it'll just increment the occurrence number by one. If it's
+     * not exists in the first place, it'll insert the word that is read and initialize with zero
+     * and then increase it to one.
+     */
+    void read(String fileName) {
+        try (BufferedReader bufferedReader = new BufferedReader(new FileReader(fileName))) {
+            for (String line; (line = bufferedReader.readLine()) != null; ) {
+                this.initTable(line);
+                if (this.isExists(line)) synchronized (lock) {
+                    Integer counter = this.value(line);
+                    this.replaceValue(line, counter, ++counter);
                 }
             }
-        } catch (IOException e) {
+        } catch (IOException e){
             e.printStackTrace();
         }
     }
