--- file1.dat
@@ -1,7 +1,0 @@
-test
-test
-test
-eer
-reee
-ree
-reee
