--- file2.dat
@@ -1,6 +1,0 @@
-scent
-sense
-scent
-sense
-sensitivity
-sensitive
